Implemented #691, except columns

+++ protected/views/project/update.php
@@ -1,24 +1,30 @@
 <?php
 declare(strict_types=1);
 
-/**
- * @var \prime\models\ar\Project $project
- * @var \prime\components\View $this
- */
-
 use app\components\ActiveForm;
 use app\components\Form;
+use kartik\select2\Select2;
 use League\ISO3166\ISO3166;
+use prime\components\View;
 use prime\helpers\Icon;
 use prime\models\ar\Permission;
+use prime\models\ar\Project;
+use prime\widgets\ButtonGroup;
 use prime\widgets\FormButtonsWidget;
 use prime\widgets\menu\ProjectTabMenu;
 use prime\widgets\Section;
 use yii\helpers\Html;
+use yii\web\User;
 use function iter\chain;
 use function iter\func\nested_index;
 use function iter\map;
 use function iter\toArrayWithKeys;
+
+/**
+ * @var Project $project
+ * @var View $this
+ * @var User $userComponent
+ */
 
 $this->title = $project->title;
 
@@ -29,20 +35,13 @@ echo ProjectTabMenu::widget([
 $this->endBlock();
 
 
-Section::begin([
-    'subject' => $project,
-    'header' => \Yii::t('app', 'Project settings'),
+Section::begin()
+    ->withSubject($project)
+    ->withHeader(\Yii::t('app', 'Project settings'));
 
-]);
 /** @var ActiveForm $form */
 $form = ActiveForm::begin([
     'method' => 'PUT',
-    "type" => ActiveForm::TYPE_HORIZONTAL,
-    'formConfig' => [
-        'showLabels' => true,
-        'defaultPlaceholder' => false,
-        'labelSpan' => 3
-    ]
 ]);
 
 echo Form::widget([
@@ -68,7 +67,7 @@ echo Form::widget([
         ],
         'country' => [
             'type' => Form::INPUT_WIDGET,
-            'widgetClass' => \kartik\select2\Select2::class,
+            'widgetClass' => Select2::class,
             'options' => [
                 'data' => toArrayWithKeys(chain(
                     ['' => \Yii::t('app', '(Not set)')],
@@ -80,52 +79,50 @@ echo Form::widget([
         'typemapAsJson' => [
             'type' => Form::INPUT_TEXTAREA,
             'options' => [
-                'rows' => 10
+                'rows' => 5
             ]
         ],
         'overridesAsJson' => [
             'type' => Form::INPUT_TEXTAREA,
             'options' => [
-                'rows' => 10
+                'rows' => 5
             ]
         ],
         FormButtonsWidget::embed([
             'buttons' => [
-                [
-                    'label' => \Yii::t('app', 'Update project'),
-                    'options' => ['class' => 'btn btn-primary'],
+                ['label' => \Yii::t('app', 'Update'), 'style' => 'primary'],
+            ],
+        ]),
+    ]
+]);
+
+ActiveForm::end();
+Section::end();
+
+if ($userComponent->can(Permission::PERMISSION_DELETE, $project)) {
+    Section::begin()
+        ->withHeader(\Yii::t('app', 'Delete project'))
+        ->forDangerousAction();
+
+    echo Html::tag('p', \Yii::t('app', 'This will permanently delete the project and all its workspaces.'));
+    echo Html::tag('p', \Yii::t('app', 'This action cannot be undone.'));
+    echo Html::tag('p', Html::tag('em', \Yii::t('app', 'Are you ABSOLUTELY SURE you wish to delete this project?')));
+
+    echo ButtonGroup::widget([
+        'buttons' => [
+            [
+                'icon' => Icon::trash(),
+                'label' => \Yii::t('app', 'Delete'),
+                'link' => ['project/delete', 'id' => $project->id],
+                'style' => 'delete',
+                'linkOptions' => [
+                    'data-method' => 'delete',
+                    'title' => \Yii::t('app', 'Delete project'),
+                    'data-confirm' => \Yii::t('app', 'Are you sure you wish to remove this project from the system?'),
                 ]
             ]
-        ])
-    ]
-]);
-$form->end();
-Section::end();
-Section::begin()
-    ->withHeader(\Yii::t('app', 'Delete project'))
-    ->forDangerousAction()
-;
-
-
-echo Html::tag('p', \Yii::t('app', 'This will permanently delete the project and all its workspaces.'));
-echo Html::tag('p', \Yii::t('app', 'This action cannot be undone.'));
-echo Html::tag('p', Html::tag('em', \Yii::t('app', 'Are you ABSOLUTELY SURE you wish to delete this project?')));
-
-echo \prime\widgets\ButtonGroup::widget([
-    'buttons' => [
-        [
-            'visible' => \Yii::$app->user->can(Permission::PERMISSION_DELETE, $project),
-            'icon' => Icon::trash(),
-            'label' => \Yii::t('app', 'Delete'),
-            'link' => ['project/delete', 'id' => $project->id],
-            'style' => 'delete',
-            'linkOptions' => [
-                'data-method' => 'delete',
-                'title' => \Yii::t('app', 'Delete project'),
-                'data-confirm' => \Yii::t('app', 'Are you sure you wish to remove this project from the system?'),
-            ]
         ]
-    ]
-]);
+    ]);
 
-Section::end();
+    Section::end();
+}
